+++ src/app.lisp
@@ -20,7 +20,8 @@
                 #:*current-app*)
   ;; Just dependencies
   (:import-from #:log)
-  
+
+  (:shadow #:restart)
   (:export
    #:defapp
    #:app
@@ -28,6 +29,8 @@
    #:get-registered-apps
    #:app-active-p
    #:start
+   #:stop
+   #:restart
    #:find-active-app
    #:get-active-apps
    #:get-prefix
@@ -183,7 +186,7 @@ called (primarily for backward compatibility"
             initargs :subclasses :slots :autostart))
        (:metaclass app-class))
      (when (find-active-app ',name :signal-error nil)
-       (restart-webapp ',name))))
+       (restart ',name))))
 
 
 (defmethod initialize-instance :after
@@ -317,10 +320,10 @@ called (primarily for backward compatibility"
 (defmethod finalize-webapp :after ((app app))
   "Shutdown Weblocks when no more apps are running."
   (unless *active-apps*
-    ;; TODO: break this tie
-    (funcall (intern "STOP-WEBLOCKS" :weblocks/server))))
+    ;; TODO: break this tie #medium
+    (uiop:symbol-call :weblocks/server :stop)))
 
-(defun stop-webapp (name)
+(defun stop (name)
   "Stops the web application"
   (log:debug "Stopping webapp" name)
   
@@ -331,8 +334,8 @@ called (primarily for backward compatibility"
       (finalize-webapp app))))
 
 
-(defun restart-webapp (name)
-  (stop-webapp name)
+(defun restart (name)
+  (stop name)
   (start name))