+++ src/debug.lisp
@@ -12,6 +12,7 @@
   (:import-from #:weblocks/request
                 #:*request*)
   (:import-from #:weblocks/variables
+                #:*catch-errors-p*
                 #:*ignore-missing-actions*)
   (:export #:*latest-session*
            #:*latest-request*
@@ -50,7 +51,8 @@ To clear, use function \(reset-last-session\).")
 (defun on (&key
              (track-latest-session t)
              (debug-actions t)
-             (track-latest-request t))
+             (track-latest-request t)
+             (invoke-debugger-on-errors t))
   ;; TODO: think about pluggable switchers to not hardcode them into this function
   (setf *on* t)
   
@@ -78,6 +80,13 @@ To clear, use function \(reset-last-session\).")
           (getf *config* :debug-actions)
           t))
 
+  (when invoke-debugger-on-errors
+    (setf *catch-errors-p*
+          nil
+          (getf *config* :invoke-debugger-on-errors)
+          t))
+
+
   (values))
 
 
@@ -92,6 +101,10 @@ To clear, use function \(reset-last-session\).")
   
   (when (getf *config* :debug-actions)
     (setf *ignore-missing-actions*
+          t))
+  
+  (when (getf *config* :invoke-debugger-on-errors)
+    (setf *catch-errors-p*
           t))
   
   (setf *config* nil)