+++ ChangeLog.rst
@@ -2,7 +2,82 @@
  ChangeLog
 ===========
 
-0.24.0 (2018-01-12)
+0.25.1 (???)
+============
+
+Old tests for widgets, removed from core framework were removed.
+
+Tests for widget's MOP methods were ported to Rove.
+
+0.25.0 (2018-01-31)
+===================
+
+Good news, everyone!
+--------------------
+
+Quickstart tutorial was fixed!
+
+Widget related changes
+----------------------
+
+Function ``render-widget`` was replaced with ``weblocks/widget:render
+:around`` method. Method implementation were moved to
+``weblocks/widgets/render-methods``.
+
+Funcall-widget's package was refactored to conform package inferred
+requirements.
+
+Actions processing
+------------------
+
+Function ``make-js-action`` was moved to ``weblocks/actions`` package.
+
+Application and server restarting
+---------------------------------
+
+Package ``weblocks/app`` now exports functions ``stop`` and ``restart``.
+Previously they were internal and were called like ``stop-webapp`` and
+``restart-webapp``.
+
+Package ``weblocks/server`` does not export ``start-weblocks`` and
+``stop-weblocks`` functions. They were replaced with just ``start`` and
+``stop``. Also, it's internal generics ``start`` and ``stop`` were
+renamed and transformed to ``start-server`` and ``stop-server``
+functions.
+
+Session initialization changes
+------------------------
+
+File ``default-application.lisp`` was removed, because now every
+application has default session initialization method.
+
+File ``default-init.lisp`` was refactored. Now it provides default
+method for ``weblocks/session:init`` generic and ``:around`` method for
+same generic, which allows end user just to return a string or a
+function from his ``init`` method.
+
+Now user can return any string or function from his method
+``weblocks/session:init`` and it will be passed to a new generic
+``weblocks/widget:create-widget-from`` to create a root widget. You can
+also define ``create-widget-from`` for you custom types.
+
+Debug
+-----
+
+Function ``weblocks/debug:on`` now turns on mode when Weblocks does not
+intercept unhandled conditions, but calls an interactive debugger.
+
+Removals
+--------
+
+* ``session-name-string-pair`` this function was removed in previous
+  releases, but didn't find it's way to the changelog, still used in
+  some older tests not ported to the Rove.
+
+* Macro ``defrender`` was removed.
+
+
+0.24.0 (2018-01-29)
 ===================
 
 All rendering code was refactored.
@@ -14,7 +89,7 @@ Stream ``*weblocks-output-stream*`` was moved to
 use it directly.
 
 Widget refactorings
-===================
+-------------------
 
 Procedure ``update-widget-tree`` was removed and not widgets can't
 change html header's tags, description, title, etc. If you need this,
@@ -24,7 +99,7 @@ Macro ``root-widget`` was removed and replaced with function
 ``weblocks.widgets.root:get``.
 
 Request level
-=============
+-------------
 
 Functions ``post-action-redirect``, ``post-render-redirect`` and
 ``initial-request-p`` were removed from ``weblocks`` package.
@@ -45,7 +120,7 @@ or deferring redirection until the end of action or rendering was
 removed.
 
 Request handler
-===============
+---------------
 
 Functions ``remove-duplicate-dirty-widgets``,
 ``update-location-hash-dependents`` and ``update-widget-tree`` were
@@ -56,14 +131,14 @@ Call to ``weblocks::update-dialog-on-request`` from
 
 
 Error handler
-=============
+-------------
 
 Generic method ``weblocks/error-handler:on-error`` now accepts two
 arguments - application object and condition.
 
 
 Application level
-=================
+-----------------
 
 All code from ``uri-parameters-slotmap.lisp`` was removed.
 
@@ -167,7 +242,7 @@ These functions were moved into the separate package
 * ``webapp-prefix`` -> ``get-prefix``;
 
 Actions and commands
-====================
+--------------------
 
 Function ``weblocks.actions:add-command`` was moved to
 ``weblocks.commands``.
@@ -178,7 +253,7 @@ Function ``weblocks:get-request-action`` was moved to
 Keyword argment ``:action`` was removed from action calls.
 
 Javascript
-==========
+----------
 
 Package ``weblocks.js`` was renamed to ``weblocks/js/base``.
 
@@ -187,7 +262,7 @@ Functions ``escape-script-tags``, ``%js`` and macroses
 package ``weblocks/js/base``.
 
 Variables
-=========
+---------
 
 These variables were moved from ``weblocks`` package to
 ``weblocks/variables``:
@@ -200,23 +275,23 @@ These variables were moved from ``weblocks`` package to
 * ``*ignore-missing-actions*``
 
 Symbols moved from :weblocks to other packages
-==============================================
+----------------------------------------------
 
 To :weblocks/widgets/dom
-------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~
 
 * ``dom-object-mixin``
 * ``dom-id``
 
 To :weblocks/utils/uri
-----------------------
+~~~~~~~~~~~~~~~~~~~~~~
 
 * ``request-uri-path``
 * ``add-get-param-to-url``
 * ``remove-parameter-from-uri``
 
 To :weblocks/linguistic/grammar
--------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 * ``pluralize``
 * ``singularize``
@@ -234,7 +309,7 @@ To :weblocks/linguistic/grammar
 * ``determine-gender``
 
 To weblocks/utils/warn
-----------------------
+~~~~~~~~~~~~~~~~~~~~~~
 
 * ``style-warn``
 * ``webapp-style-warning`` renamed to ``style-warning``.
@@ -242,7 +317,7 @@ To weblocks/utils/warn
 * ``misunderstood-action``
 
 To weblocks/actions
--------------------
+~~~~~~~~~~~~~~~~~~~
 
 * ``function-or-action->action``
 * ``make-action``
@@ -250,7 +325,7 @@ To weblocks/actions
 
 
 Removals
-========
+--------
 
 To make Weblocks core smaller, many files were removed: ``views``, ``widgets``,
 ``html-parts``, ``utilities``.
@@ -267,7 +342,7 @@ Variable ``*dirty-widgets*`` was removed along with
 
 
 Dependencies
-============
+------------
 
 Rendering of remote (non cached) dependencies was fixed.